+++ src/state/mod.rs
@@ -7,13 +7,63 @@ use pallas::{
     },
 };
 use std::collections::{HashMap, HashSet};
+use thiserror::Error;
 
 use crate::ledger::*;
 
 pub mod redb;
 
+#[derive(Debug, Error)]
+pub enum LedgerError {
+    #[error("broken invariant")]
+    BrokenInvariant(#[source] BrokenInvariant),
+
+    #[error("storage error")]
+    StorageError(#[source] ::redb::Error),
+
+    #[error("address decoding error")]
+    AddressDecoding(pallas::ledger::addresses::Error),
+
+    #[error("query not supported")]
+    QueryNotSupported,
+
+    #[error("invalid store version")]
+    InvalidStoreVersion,
+}
+
+impl From<::redb::TableError> for LedgerError {
+    fn from(value: ::redb::TableError) -> Self {
+        Self::StorageError(value.into())
+    }
+}
+
+impl From<::redb::CommitError> for LedgerError {
+    fn from(value: ::redb::CommitError) -> Self {
+        Self::StorageError(value.into())
+    }
+}
+
+impl From<::redb::StorageError> for LedgerError {
+    fn from(value: ::redb::StorageError) -> Self {
+        Self::StorageError(value.into())
+    }
+}
+
+impl From<::redb::TransactionError> for LedgerError {
+    fn from(value: ::redb::TransactionError) -> Self {
+        Self::StorageError(value.into())
+    }
+}
+
+impl From<pallas::ledger::addresses::Error> for LedgerError {
+    fn from(value: pallas::ledger::addresses::Error) -> Self {
+        Self::AddressDecoding(value)
+    }
+}
+
 /// A persistent store for ledger state
 #[derive(Clone)]
+#[non_exhaustive]
 pub enum LedgerStore {
     Redb(redb::LedgerStore),
 }
@@ -82,6 +132,18 @@ impl LedgerStore {
     pub fn finalize(&mut self, until: BlockSlot) -> Result<(), LedgerError> {
         match self {
             LedgerStore::Redb(x) => x.finalize(until),
+        }
+    }
+
+    pub fn upgrade(self) -> Result<Self, LedgerError> {
+        match self {
+            LedgerStore::Redb(x) => Ok(LedgerStore::Redb(x.upgrade()?)),
+        }
+    }
+
+    pub fn copy(&self, target: &Self) -> Result<(), LedgerError> {
+        match (self, target) {
+            (Self::Redb(x), Self::Redb(target)) => x.copy(target),
         }
     }
 }
@@ -153,7 +215,7 @@ pub fn load_slice_for_block(
     Ok(LedgerSlice { resolved_inputs })
 }
 
-pub fn import_block_batch(
+pub fn apply_block_batch(
     blocks: &[MultiEraBlock],
     store: &mut LedgerStore,
     byron: &byron::GenesisFile,