+++ CHANGELOG.md
@@ -2,6 +2,34 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.13.1] - 2024-07-19
+
+### 🐛 Bug Fixes
+
+- *(bootstrap)* Save ledger to disk before attempting upgrade (#310)
+
+### 🚜 Refactor
+
+- *(state)* Treat address parsing as a fallible operation (#309)
+
+### ⚙️ Miscellaneous Tasks
+
+- Use Pallas edge with long address fix (#311)
+
+## [0.13.0] - 2024-07-18
+
+### 🚀 Features
+
+- Bootstrap in-memory and with defered indexes (#308)
+
+### 🐛 Bug Fixes
+
+- *(state)* Use cursor value to decide if db is empty (#302)
+
+### ⚙️ Miscellaneous Tasks
+
+- Release dolos version 0.13.0
+
 ## [0.12.0] - 2024-07-16
 
 ### 🚀 Features
@@ -13,6 +41,7 @@ All notable changes to this project will be documented in this file.
 
 - Update Pallas with u5c spec v0.7.0 (#300)
 - Update Pallas to v0.29 (#301)
+- Release dolos version 0.12.0
 
 ## [0.11.1] - 2024-07-14